Extracted styles into their own files, fixed issue where nodes were being rendered off the graph
