+++ src/components/DagreGraph/GraphRoot.styles.ts
@@ -1,0 +1,12 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+export default makeStyles(
+  {
+    root: {
+      position: 'relative',
+      minWidth: '100%',
+      minHeight: '100%'
+    }
+  },
+  { name: 'DagreGraph-GraphRoot' }
+);
